+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+-
+
+## 4.39.0
 - Bundle BitBox02 firmware version v9.15.0
 - Display the wallet root fingerprint in the account info and device settings